requirements.txt
python-dotenv
aiofiles
faiss-cpu
langchain
langchain_community
langchain_openai
langchain-huggingface
langchain-ollama
uvicorn
fastapi